Revert "Изменение подхода к расчету приоритета тикеров."
